Add support for JSON output
Fixes #21

TODO: If I were doing this from scratch, I would probably implement JSON as the default output and then have a wrapper that pretty-prints the output as text.  As things stand, there are two distinct implementations that share a lot of helper functions.  The biggest difference between the two is that JSON nodes use explicit `children` properties (i.e. follow a schema), whereas text output nodes follow treeify's structure (i.e. children are dynamically generated properties on parent node).
